+++ pkg/controller/atlasstream/transitions.go
@@ -100,6 +100,7 @@ func (r *InstanceReconciler) skip(ctx context.Context, log *zap.SugaredLogger, s
 
 // transitions back to pending state setting an terminate state
 func (r *InstanceReconciler) invalidate(invalid workflow.Result) (ctrl.Result, error) {
+	// note: ValidateResourceVersion already set the state so we don't have to do it here.
 	r.Log.Debugf("AtlasStreamInstance is invalid: %v", invalid)
 	return invalid.ReconcileResult(), nil
 }
